# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LidRadar API
  version: 1.0.0
  description: >-
    Целевой контракт MVP. Пути без расширения x-lidradar-runtime-status уже
    подключены к рабочему API. Значение planned означает, что контракт
    зафиксирован для будущего этапа, но маршрут ещё не доступен в cmd/api.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current LidRadar API origin
security:
  - sessionCookie: []
paths:
...
  /api/v1/radar:
    get:
      summary: Получить сводку Radar
      operationId: getRadarSummary
      parameters:
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { name: severity, in: query, schema: { $ref: '#/components/schemas/RiskSeverity' } }
        - { name: riskType, in: query, schema: { $ref: '#/components/schemas/RiskType' } }
      responses:
        '200':
          description: Сводка организации. Денежные значения представлены точными десятичными строками в основной валюте организации.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RadarSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/risks:
    get:
      summary: Получить список рисков организации
      operationId: listRisks
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/RiskStatus' } }
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { name: severity, in: query, schema: { $ref: '#/components/schemas/RiskSeverity' } }
        - { name: riskType, in: query, schema: { $ref: '#/components/schemas/RiskType' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Риски в порядке серверного приоритета и непрозрачный курсор продолжения.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/RiskDetail' } }
                  nextCursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/risks/{riskId}:
    parameters: [{ $ref: '#/components/parameters/RiskId' }]
    get:
      summary: Получить детали риска
      operationId: getRisk
      responses:
        '200':
          description: Риск и связанные данные, уже созданные владеющими модулями.
          content: { application/json: { schema: { $ref: '#/components/schemas/RiskDetail' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risks/{riskId}/acknowledge:
    parameters: [{ $ref: '#/components/parameters/RiskId' }]
    post:
      summary: Подтвердить просмотр риска
      operationId: acknowledgeRisk
      responses:
        '200': { description: Текущее состояние риска; повтор команды идемпотентен., content: { application/json: { schema: { $ref: '#/components/schemas/Risk' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risks/{riskId}/resolve:
    parameters: [{ $ref: '#/components/parameters/RiskId' }]
    post:
      summary: Закрыть риск
      operationId: resolveRisk
      responses:
        '200': { description: Текущее состояние риска; повтор команды идемпотентен., content: { application/json: { schema: { $ref: '#/components/schemas/Risk' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risks/precision:
    parameters: [{ $ref: '#/components/parameters/TenantId' }]
    get:
      summary: Точность рисков по типам за окно обнаружения
      description: Каждый риск учитывается один раз по последнему вердикту (LR-BE-RM-019); coverageRate — доля рисков окна с обратной связью, reliable — покрытие не ниже minimumCoverage. Требует analytics.read.
      operationId: getRiskPrecision
      parameters:
        - { name: from, in: query, required: false, description: 'Начало окна обнаружения (RFC 3339); по умолчанию — с начала.', schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, description: 'Конец окна обнаружения (RFC 3339, исключительно); по умолчанию — сейчас.', schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Отчёт по пяти типам риска.
          content: { application/json: { schema: { $ref: '#/components/schemas/RiskPrecisionReport' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/risks/{riskId}/feedback:
    parameters: [{ $ref: '#/components/parameters/TenantId' }, { $ref: '#/components/parameters/RiskId' }]
    post:
      summary: Записать вердикт по риску
      description: Append-only факт со снимком риска и сделки на момент вердикта. FALSE_POSITIVE закрывает активный риск как ложное срабатывание; причина NOT_A_LEAD дополнительно закрывает сделку как LOST. Требует risks.manage.
      operationId: recordRiskFeedback
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskFeedbackRequest' } } }
      responses:
        '201':
          description: Вердикт сохранён.
          content: { application/json: { schema: { $ref: '#/components/schemas/RiskFeedback' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/events:
    get:
      summary: Получать сигналы об изменениях организации
      operationId: streamInvalidations
      description: Временные сигналы в границах организации без бизнес-состояния; после сигнала клиент перечитывает авторитетный REST-ресурс.
      responses:
        '200': { description: "Поток text/event-stream с событиями risk.changed, risk.acknowledged, risk.resolved и risk.false_positive." }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/opportunities/{opportunityId}/revenue:
    parameters:
      - { $ref: '#/components/parameters/TenantId' }
      - { $ref: '#/components/parameters/OpportunityId' }
      - { $ref: '#/components/parameters/IdempotencyKey' }
    post:
      summary: Подтвердить выручку коммерческой возможности
      description: Событие, его единственная формальная атрибуция, ключ идемпотентности и аудит записываются одной транзакцией. Для RECOVERED обязательна 30-дневная цепочка Risk → Action → Outcome той же организации и возможности.
      operationId: confirmRevenue
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConfirmRevenueRequest' } } } }
      responses:
        '200': { description: Ранее сохранённое подтверждение при точном повторе., content: { application/json: { schema: { $ref: '#/components/schemas/RevenueConfirmation' } } } }
        '201': { description: Подтверждённая выручка и её единственная атрибуция., content: { application/json: { schema: { $ref: '#/components/schemas/RevenueConfirmation' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Ключ идемпотентности использован с другим содержимым., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/revenue/confirmed-recovered:
    parameters:
      - { $ref: '#/components/parameters/TenantId' }
    get:
      summary: Получить подтверждённую возвращённую выручку
      description: Суммируются только CONFIRMED-события с формальной атрибуцией RECOVERED; валюты никогда не смешиваются.
      operationId: getConfirmedRecoveredRevenue
      parameters: [{ name: currency, in: query, required: true, schema: { type: string, pattern: '^[A-Z]{3}$' } }]
      responses:
        '200': { description: Формальная возвращённая выручка в одной валюте., content: { application/json: { schema: { $ref: '#/components/schemas/Money' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lidradar_session
    aiNodeBearer:
      type: http
      scheme: bearer
      bearerFormat: LidRadar AI node secret
  parameters:
    TenantId: { name: X-Tenant-ID, in: header, required: true, description: Explicit Organization tenant selected from /auth/me memberships., schema: { type: string, format: uuid } }
...
    OpportunityId: { name: opportunityId, in: path, required: true, schema: { type: string, format: uuid } }
...
    RiskId: { name: riskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
...
  responses:
    BadRequest: { description: 'Request body, tenant selector, or domain value is invalid.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: 'Opaque session cookie is missing, expired, revoked, or invalid.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Forbidden: { description: Request origin is not trusted or Membership does not grant the required named permission., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource does not exist in the authenticated tenant, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties: { error: { type: object, required: [code, message, traceId], properties: { code: { type: string }, message: { type: string }, details: { type: object, additionalProperties: true }, traceId: { type: string } } } }
...
    OpportunityStage:
      type: string
      enum: [NEW, ENGAGED, QUALIFYING, PRICE_SENT, WAITING_CUSTOMER, WAITING_BUSINESS, BOOKING_INTENT, BOOKED, WON, LOST, ARCHIVED]
...
    RiskStatus: { type: string, enum: [OPEN, ACKNOWLEDGED, ACTED, RESOLVED, FALSE_POSITIVE, IGNORED, EXPIRED] }
    RiskSeverity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    RiskType: { type: string, enum: [NO_RESPONSE, BOOKING_NOT_CONFIRMED, PROMISE_NOT_FULFILLED, CUSTOMER_SILENT_AFTER_PRICE, FOLLOW_UP_CANDIDATE] }
...
    Money:
      type: object
      required: [amount, currency]
      properties: { amount: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }, currency: { type: string, pattern: '^[A-Z]{3}$' } }
    ConfirmRevenueRequest:
      type: object
      additionalProperties: false
      required: [amount, currency, attributionType]
      properties:
        amount: { type: string, pattern: '^[0-9]{1,12}(\.[0-9]{1,2})?$' }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
        attributionType: { type: string, enum: [RECOVERED, ORGANIC, UNKNOWN] }
        riskId: { type: string, format: uuid }
        actionId: { type: string, format: uuid }
        outcomeId: { type: string, format: uuid }
    RevenueEvent:
      type: object
      additionalProperties: false
      required: [id, opportunityId, amount, currency, status, source, confirmedBy, confirmedAt]
      properties: { id: { type: string, format: uuid }, opportunityId: { type: string, format: uuid }, amount: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }, currency: { type: string, pattern: '^[A-Z]{3}$' }, status: { type: string, enum: [CONFIRMED] }, source: { type: string, enum: [USER_CONFIRMED] }, confirmedBy: { type: string, format: uuid }, confirmedAt: { type: string, format: date-time } }
    RevenueAttribution:
      type: object
      additionalProperties: false
      required: [id, revenueEventId, opportunityId, type, createdAt]
      properties: { id: { type: string, format: uuid }, revenueEventId: { type: string, format: uuid }, opportunityId: { type: string, format: uuid }, type: { type: string, enum: [RECOVERED, ORGANIC, UNKNOWN] }, riskId: { type: string, format: uuid }, actionId: { type: string, format: uuid }, outcomeId: { type: string, format: uuid }, createdAt: { type: string, format: date-time } }
    RevenueConfirmation:
      type: object
      additionalProperties: false
      required: [revenue, attribution]
      properties: { revenue: { $ref: '#/components/schemas/RevenueEvent' }, attribution: { $ref: '#/components/schemas/RevenueAttribution' } }
    Risk:
      type: object
      additionalProperties: false
      required: [id, opportunityId, locationId, type, severity, status, source, policyVersion, triggerMessageId, reasonCode, reason, detectedAt, dueAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        opportunityId: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/RiskType' }
        severity: { $ref: '#/components/schemas/RiskSeverity' }
        status: { $ref: '#/components/schemas/RiskStatus' }
        source: { type: string, enum: [RULE, HYBRID] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        aiRunId: { type: string, format: uuid }
        policyVersion: { type: string }
        triggerMessageId: { type: string, format: uuid }
        reasonCode: { type: string, pattern: '^[A-Z][A-Z0-9_]{0,99}$' }
        reason: { type: string }
        detectedAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        acknowledgedAt: { type: string, format: date-time }
        actedAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time }
    RadarOpportunity:
      type: object
      additionalProperties: false
      required: [id, stage, locationId, potentialRevenue, currency]
      properties:
        id: { type: string, format: uuid }
        stage: { $ref: '#/components/schemas/OpportunityStage' }
        locationId: { type: string, format: uuid }
        potentialRevenue: { type: [string, 'null'], pattern: '^[0-9]+\.[0-9]{2}$' }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
    RadarConversation:
      type: object
      additionalProperties: false
      required: [id, contactId]
      properties:
        id: { type: string, format: uuid }
        contactId: { type: string, format: uuid }
    RadarRecommendation:
      type: object
      additionalProperties: false
      required: [id, text]
      properties: { id: { type: string, format: uuid }, text: { type: string } }
    RadarAction:
      type: object
      additionalProperties: false
      required: [id, type, createdAt]
      properties: { id: { type: string, format: uuid }, type: { type: string }, createdAt: { type: string, format: date-time } }
    RadarOutcome:
      type: object
      additionalProperties: false
      required: [id, type, createdAt]
      properties: { id: { type: string, format: uuid }, type: { type: string }, createdAt: { type: string, format: date-time } }
    RadarRevenue:
      type: object
      additionalProperties: false
      required: [currency, potential, confirmedRecovered]
      properties:
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        potential: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }
        confirmedRecovered: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }
    RiskDetail:
      type: object
      additionalProperties: false
      required: [risk, opportunity, conversation, actions]
      properties:
        risk: { $ref: '#/components/schemas/Risk' }
        opportunity: { $ref: '#/components/schemas/RadarOpportunity' }
        conversation: { $ref: '#/components/schemas/RadarConversation' }
        recommendation: { $ref: '#/components/schemas/RadarRecommendation' }
        actions: { type: array, items: { $ref: '#/components/schemas/RadarAction' } }
        outcome: { $ref: '#/components/schemas/RadarOutcome' }
        revenue: { $ref: '#/components/schemas/RadarRevenue' }
    RiskVerdict: { type: string, enum: [TRUE_POSITIVE, FALSE_POSITIVE] }
    RiskFeedbackReason: { type: string, enum: [CUSTOMER_ALREADY_BOOKED, CUSTOMER_ALREADY_ANSWERED, NOT_A_LEAD, CUSTOMER_REJECTED, WRONG_INTERPRETATION, OTHER] }
    RiskFeedbackRequest:
      type: object
      additionalProperties: false
      required: [verdict]
      properties:
        verdict: { $ref: '#/components/schemas/RiskVerdict' }
        reason: { $ref: '#/components/schemas/RiskFeedbackReason', description: Обязательна для FALSE_POSITIVE. }
        note: { type: string, maxLength: 1000 }
    RiskFeedbackContext:
      type: object
      additionalProperties: false
      required: [type, severity, status, source, policyVersion, triggerMessageId, opportunityStage, detectedAt]
      properties:
        type: { $ref: '#/components/schemas/RiskType' }
        severity: { $ref: '#/components/schemas/RiskSeverity' }
        status: { $ref: '#/components/schemas/RiskStatus' }
        source: { type: string, enum: [RULE, HYBRID, MANUAL] }
        policyVersion: { type: string }
        aiRunId: { type: string, format: uuid }
        triggerMessageId: { type: string, format: uuid }
        opportunityStage: { $ref: '#/components/schemas/OpportunityStage' }
        detectedAt: { type: string, format: date-time }
    RiskFeedback:
      type: object
      additionalProperties: false
      required: [id, riskId, opportunityId, actorId, verdict, note, context, datasetEligible, createdAt]
      properties:
        id: { type: string, format: uuid }
        riskId: { type: string, format: uuid }
        opportunityId: { type: string, format: uuid }
        actorId: { type: string, format: uuid }
        verdict: { $ref: '#/components/schemas/RiskVerdict' }
        reason: { $ref: '#/components/schemas/RiskFeedbackReason' }
        note: { type: string }
        context: { $ref: '#/components/schemas/RiskFeedbackContext' }
        datasetEligible: { type: boolean, description: Действовало ли явное ML-согласие организации в момент записи (§70). }
        createdAt: { type: string, format: date-time }
    RiskPrecisionItem:
      type: object
      additionalProperties: false
      required: [riskType, totalRisks, withFeedback, truePositives, falsePositives, precision, falsePositiveRate, coverageRate, reliable]
      properties:
        riskType: { $ref: '#/components/schemas/RiskType' }
        totalRisks: { type: integer, minimum: 0 }
        withFeedback: { type: integer, minimum: 0 }
        truePositives: { type: integer, minimum: 0 }
        falsePositives: { type: integer, minimum: 0 }
        precision: { type: [number, 'null'], minimum: 0, maximum: 1 }
        falsePositiveRate: { type: [number, 'null'], minimum: 0, maximum: 1 }
        coverageRate: { type: number, minimum: 0, maximum: 1 }
        reliable: { type: boolean }
    RiskPrecisionReport:
      type: object
      additionalProperties: false
      required: [from, to, minimumCoverage, items]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        minimumCoverage: { type: number }
        items: { type: array, items: { $ref: '#/components/schemas/RiskPrecisionItem' } }
...
    RadarSummary:
      type: object
      required: [openRisks, criticalRisks, potentialRevenue, confirmedRecoveredRevenue]
      properties:
        openRisks: { type: integer, minimum: 0 }
        criticalRisks: { type: integer, minimum: 0 }
        potentialRevenue: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }
        confirmedRecoveredRevenue: { type: string, pattern: '^[0-9]+\.[0-9]{2}$' }